+++ aqchem/__init__.py
@@ -2,5 +2,5 @@
 aqchem is a Python package useful for (physical) chemistry.
 """
 
-from .release import __version__
+from ._release import __version__
 __url__ = "https://github.com/bjodah/aqchem"
