+++ aqchem/tests/test_equilibria.py
@@ -3,7 +3,7 @@ from scipy.optimize import fsolve
 
 from ..equilibria import (
     equilibrium_quotient, equilibrium_residual, get_rc_interval,
-    solve_equilibrium
+    solve_equilibrium, EqSystem, prodexp
 )
 
 
@@ -30,7 +30,7 @@ def test_get_rc_interval():
     assert abs(limits[1] - upper) < 1e-14
 
 
-def test_solve_equilibrium():
+def test_solve_equilibrium_1():
     c = np.array((13., 11, 17))
     stoich = np.array((-2, 3, -4))
     K = 3.14
@@ -40,3 +40,17 @@ def test_solve_equilibrium():
             11 + 3*x)**3 * (17 - 4*x)**-4 - K
     assert np.allclose(solve_equilibrium(c, stoich, K),
                        c + stoich*fsolve(f, 3.48))
+
+def test_solve_equilibrium_2():
+    c = np.array([  1.7000e-03,   3.0000e+06,   3.0000e+06,   9.7000e+07,   5.5500e+09])
+    stoich = (1, 1, 0, 0, -1)
+    K = 1e-14
+
+    def f(x):
+        return prodexp(c, stoich) - K
+    assert np.allclose(solve_equilibrium(c, stoich, K),
+                       c + stoich*fsolve(f, 3.48))
+
+
+def test_EqSystem():
+    pass
